+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/audio/TranscriptionRequest.kt
@@ -1,6 +1,5 @@
 package com.aallam.openai.api.audio
 
-import com.aallam.openai.api.BetaOpenAI
 import com.aallam.openai.api.OpenAIDsl
 import com.aallam.openai.api.file.FileSource
 import com.aallam.openai.api.model.ModelId
@@ -43,6 +42,14 @@ public class TranscriptionRequest(
      * [ISO-639-1](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) format will improve accuracy and latency.
      */
     public val language: String? = null,
+
+    /**
+     * The timestamp granularities to populate for this transcription.
+     * [responseFormat] must be set [AudioResponseFormat.VerboseJson] to use timestamp granularities.
+     * Either or both of these options are supported: [TimestampGranularity.Word], or [TimestampGranularity.Segment].
+     * Note: There is no additional latency for segment timestamps, but generating word timestamps incurs additional latency.
+     */
+    public val timestampGranularities: List<TimestampGranularity>? = null,
 )
 
 /**
@@ -91,6 +98,14 @@ public class TranscriptionRequestBuilder {
     public var language: String? = null
 
     /**
+     * The timestamp granularities to populate for this transcription.
+     * responseFormat must be set verbose_json to use timestamp granularities.
+     * Either or both of these options are supported: word, or segment.
+     * Note: There is no additional latency for segment timestamps, but generating word timestamps incurs additional latency.
+     */
+    public var timestampGranularities: List<TimestampGranularity>? = null
+
+    /**
      * Builder of [TranscriptionRequest] instances.
      */
     public fun build(): TranscriptionRequest = TranscriptionRequest(
@@ -100,5 +115,6 @@ public class TranscriptionRequestBuilder {
         responseFormat = responseFormat,
         temperature = temperature,
         language = language,
+        timestampGranularities = timestampGranularities,
     )
 }