+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+### Added
+
+- **Audio**: add `timestampGranularities` (thanks @mxwell)
+
+### Fixed
+
+- **Core**: nullable `OpenAIErrorDetails` fields (#315)
+- **Messages**: nullable field `MessageContent.Image#fileId` (#313)
+
 ## 3.7.0
 
 ### Added